Extract stat checking fn to utils
