+++ v3/src/js/views/browse/CorsBiddingStatsTableControl.jsx
@@ -7,8 +7,9 @@ import { selectNewStudent, selectFaculty } from 'actions/settings';
 import AccountSelect from 'views/components/AccountSelect';
 import FacultySelect from 'views/components/FacultySelect';
 import NewStudentSelect from 'views/components/NewStudentSelect';
-import { AccountType, BiddingStat, Faculty } from 'types/modules';
+import { AccountType, BiddingStat, Faculty, Student } from 'types/modules';
 import ButtonGroupSelector from 'views/components/ButtonGroupSelector';
+import { isStatRelevantForStudent } from 'utils/cors';
 import CorsBiddingStatsTable from './CorsBiddingStatsTable';
 
 type Props = {
@@ -18,46 +19,6 @@ type Props = {
   selectNewStudent: Function,
   stats: Array<BiddingStat>,
 };
-
-type Student = {
-  newStudent: boolean,
-  faculty: Faculty,
-  accountType: AccountType,
-}
-
-const sameFaculty = (stat: BiddingStat, student: Student): boolean => (stat.Faculty === student.faculty);
-const isNew = student => (student.newStudent);
-const PROGRAMME = 'P';
-const GENERAL = 'G';
-const isProgrammeAccount = student => (student.accountType === PROGRAMME);
-const isGeneralAccount = student => (student.accountType === GENERAL);
-
-function isStatRelevantForStudent(stat, student) {
-  switch (stat.StudentAcctType) {
-    case 'Returning Students [P]':
-      return sameFaculty(stat, student) && isProgrammeAccount(student) && !isNew(student);
-    case 'New Students [P]':
-      return sameFaculty(stat, student) && isProgrammeAccount(student) && isNew(student);
-    case 'NUS Students [P]':
-      return sameFaculty(stat, student) && isProgrammeAccount(student);
-    case 'Returning Students and New Students [P]':
-      return sameFaculty(stat, student) && isProgrammeAccount(student);
-    case 'NUS Students [G]':
-      return isGeneralAccount(student);
-    case 'Returning Students [P] and NUS Students [G]':
-      return (sameFaculty(stat, student) && isProgrammeAccount(student) && !isNew(student))
-        || (!sameFaculty(stat, student) && isGeneralAccount(student));
-    case 'NUS Students [P, G]':
-      return (sameFaculty(stat, student) && isProgrammeAccount(student))
-        || (!sameFaculty(stat, student) && isGeneralAccount(student));
-    case 'Reserved for [G] in later round':
-      return !sameFaculty(stat, student) && isGeneralAccount(student);
-    case 'Not Available for [G]':
-      return sameFaculty(stat, student) && isProgrammeAccount(student);
-    default:
-      throw Error(`unknown StudentAcctType ${stat.StudentAcctType}`);
-  }
-}
 
 class CorsBiddingStatsTableControl extends Component {
   constructor(props: Props) {
